pyproject.toml
[tool.poetry]
name = "sobitie"
version = "0.1.0"
description = "sobitie_bot"
authors = [
    "Владислав Бойко", "Никита Цыбин", "Андрей Укладчиков",
    "Екатерина Каричева", "Кирилл Резник", "внесите себя"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.21.1"
python-telegram-bot = "^20.1"
emoji = "^2.2.0"

djangorestframework = "^3.14.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
